requirements.txt
opencv-python
onnxruntime-gpu==1.9.0
torch
numpy
tqdm
selenium==4.1.5
webdriver-manager==3.5.4
